requirements.txt
asgiref==3.2.10
asn1crypto==1.4.0
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
codecov==2.1.8
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==3.3.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-crequest==2018.5.11
django-filter==2.3.0
django-import-export==2.3.0
django-nose==1.4.6
django-reversion==3.0.7
djangorestframework==3.11.1
djangorestframework-csv==2.1.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
et-xmlfile==1.0.1
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==3.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.4
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rfc3339-validator==0.1.2
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.10
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
